Seperate store creation and initialisation fucntiosn to utils file
